requirements.txt
fastapi
huggingface_hub
llama-cpp-python
openai
pydantic
requests
uvicorn